# ports/openapi.yml
openapi: "3.0.0"
info:
  version: 1.0.0
  title: Task Manager
servers:
  - url: /api

paths:
  /users:
    post:
      operationId: addUser
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUser"
      responses:
        "201":
          description: todo
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks:
    get:
      operationId: getTasks
      responses:
        "200":
          description: todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createTask
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostTask"
      responses:
        "201":
          description: todo
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/status:
    put:
      operationId: changeTaskStatus
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutTaskStatus"
      responses:
        "200":
          description: Task status successfully updated
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}:
    patch:
      operationId: updateTaskTitle
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: todo
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchTaskTitle"
      responses:
        "200":
          description: Task title updated
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /tasks/{taskId}/assign/{assigneeId}:
    put:
      operationId: assignTask
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
        - name: assigneeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Task successfully assigned
        default:
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          example: 500
          description: Error code
        message:
          type: string
          description: Error message
    Task:
      type: object
      required:
        [uuid, title, status, createdBy, assignedTo, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: Mariusz Pudzianowski
        status:
          type: string
          example: todo
        createdBy:
          type: string
          format: uuid
        assignedTo:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    PostUser:
      type: object
      required: [role]
      properties:
        role:
          type: string
          example: "employer"
    PutTaskStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string
          example: "todo"
    PatchTaskTitle:
      type: object
      required: [title]
      properties:
        title:
          type: string
          example: "Updated task title"
    PostTask:
      type: object
      required: [title, creator]
      properties:
        title:
          type: string
          example: "let's do leg day!"
        creator:
          type: string
          format: uuid