# ports/openapi_types.gen.go
// Package ports provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package ports

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Error defines model for Error.
type Error struct {
	// Code Error code
	Code int32 `json:"code"`

	// Message Error message
	Message string `json:"message"`
}

// PatchTaskTitle defines model for PatchTaskTitle.
type PatchTaskTitle struct {
	Title string `json:"title"`
}

// PostTask defines model for PostTask.
type PostTask struct {
	Creator openapi_types.UUID `json:"creator"`
	Title   string             `json:"title"`
}

// PostUser defines model for PostUser.
type PostUser struct {
	Role string `json:"role"`
}

// PutTaskStatus defines model for PutTaskStatus.
type PutTaskStatus struct {
	Status string `json:"status"`
}

// Task defines model for Task.
type Task struct {
	AssignedTo openapi_types.UUID  `json:"assignedTo"`
	CreatedAt  time.Time           `json:"createdAt"`
	CreatedBy  openapi_types.UUID  `json:"createdBy"`
	Id         *openapi_types.UUID `json:"id,omitempty"`
	Status     string              `json:"status"`
	Title      string              `json:"title"`
	UpdatedAt  time.Time           `json:"updatedAt"`
}

// CreateTaskJSONRequestBody defines body for CreateTask for application/json ContentType.
type CreateTaskJSONRequestBody = PostTask

// UpdateTaskTitleJSONRequestBody defines body for UpdateTaskTitle for application/json ContentType.
type UpdateTaskTitleJSONRequestBody = PatchTaskTitle

// ChangeTaskStatusJSONRequestBody defines body for ChangeTaskStatus for application/json ContentType.
type ChangeTaskStatusJSONRequestBody = PutTaskStatus

// AddUserJSONRequestBody defines body for AddUser for application/json ContentType.
type AddUserJSONRequestBody = PostUser